move h5_utils into shared_utils and delete file
