+++ convert_nsidc_h5_to_geotiff.py
@@ -9,7 +9,7 @@ import numpy as np
 from pathlib import Path
 import pyproj
 
-from h5_utils import (
+from shared_utils import (
     get_attrs_from_h5,
     create_proj_from_viirs_snow_h5,
     convert_data_array_to_geotiff,
